Object loading and basic light shader

+++ VS2017/Source/tests/TestTexture2D.cpp
@@ -32,7 +32,7 @@ namespace test {
 			0.0f, 1.0f
 		};
 
-		unsigned int indices[] = {
+		unsigned short indices[] = {
 			0, 1, 2,
 			2, 3, 0
 		};
@@ -55,7 +55,7 @@ namespace test {
 		layout2.Push<float>(2);
 		m_VAO->AddBuffer(*m_VertexBuffer2, layout2);
 
-		m_IndexBuffer = std::make_unique<IndexBuffer>(indices, sizeof(indices)/sizeof(unsigned int));
+		m_IndexBuffer = std::make_unique<IndexBuffer>(indices, sizeof(indices)/sizeof(unsigned short));
 
 		m_Proj = glm::ortho(0.0f, 960.0f, 0.0f, 540.0f, -1.0f, 1.0f);
 		m_View = glm::translate(glm::mat4(1.0f), glm::vec3(0.0f, 0.0f, 0.0f));
